pyproject.toml
[project]
name = "baseball-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "gunicorn>=23.0.0",
    "line-bot-sdk>=3.17.1",
    "pydantic>=2.11.3",
    "python-dotenv>=1.1.0",
    "ruff>=0.11.7",
    "uvicorn[standard]>=0.34.2",
    "langchain>=0.3.24",
    "langchain-community>=0.3.22",
    "langchain-core>=0.3.56",
    "langchain-openai>=0.3.14",
    "openai>=1.76.0",
    "requests>=2.32.3",
    "beautifulsoup4>=4.13.4",
]

[tool.ruff]
line-length = 120
